Create data models for ViewModel
