+++ app/src/main/java/com/rockthevote/grommet/ui/registration/NewRegistrantFragment.java
@@ -168,8 +168,11 @@ public class NewRegistrantFragment extends BaseRegistrationFragment implements
         Observable<Boolean> previousNameObs = previousName.verify()
                 .flatMap(val -> Observable.just(nameChanged.isChecked() ? val : true));
 
-        return Observable.zip(validator.validate(), name.verify(),
-                previousNameObs, Observable.just(isEighteenByDeadline()),
+        return Observable.zip(
+                validator.validate(),
+                name.verify(),
+                previousNameObs,
+                Observable.just(isEighteenByDeadline()),
                 (validator, name, prevName, birthday)
                         -> validator && name && prevName && birthday);
     }
