+++ app/src/main/java/com/rockthevote/grommet/ui/registration/RegistrationViewModel.kt
@@ -1,0 +1,9 @@
+package com.rockthevote.grommet.ui.registration
+
+import androidx.lifecycle.ViewModel
+
+class RegistrationViewModel : ViewModel() {
+
+
+}
+
