+++ app/src/main/java/com/rockthevote/grommet/ui/registration/address/Address.kt
@@ -1,0 +1,14 @@
+package com.rockthevote.grommet.ui.registration.address
+
+data class Address(
+    // Mandatory
+    val streetAddress: String,
+    val city: String,
+    val state: String,
+    val zipCode: String,
+    val county: String,
+    // Optional
+    val streetAddressTwo: String?,
+    val unitType: String?,
+    val unitNumber: String?
+)
