+++ app/src/main/java/com/rockthevote/grommet/ui/registration/address/AddressRegistrationData.kt
@@ -1,0 +1,12 @@
+package com.rockthevote.grommet.ui.registration.address
+
+data class AddressRegistrationData(
+    // Mandatory
+    val registrationAddress: Address,
+    val isMailingAddressDifferent: Boolean,
+    val hasPreviousAddress: Boolean,
+    // Optional
+    val mailingAddress: Address?,
+    val previousAddress: Address?
+)
+
