+++ app/src/main/java/com/rockthevote/grommet/ui/registration/assistance/AssistanceRegistrationData.kt
@@ -1,0 +1,14 @@
+package com.rockthevote.grommet.ui.registration.assistance
+
+import com.rockthevote.grommet.ui.registration.address.Address
+import com.rockthevote.grommet.ui.registration.name.PersonName
+
+data class AssistanceRegistrationData(
+    // Mandatory
+    val hasSomeoneAssisted: Boolean,
+    // Optional
+    val helperName: PersonName?,
+    val helperAddress: Address?,
+    val helperPhone: String?,
+    val hasConfirmedTerms: Boolean?
+)
