+++ app/src/main/java/com/rockthevote/grommet/ui/registration/review/ReviewRegistrationData.kt
@@ -1,0 +1,24 @@
+package com.rockthevote.grommet.ui.registration.review
+
+data class ReviewRegistrationData(
+    val hasReviewedAndConfirmedInfo: Boolean,
+    val signature: ByteArray
+) {
+    override fun equals(other: Any?): Boolean {
+        if (this === other) return true
+        if (javaClass != other?.javaClass) return false
+
+        other as ReviewRegistrationData
+
+        if (hasReviewedAndConfirmedInfo != other.hasReviewedAndConfirmedInfo) return false
+        if (!signature.contentEquals(other.signature)) return false
+
+        return true
+    }
+
+    override fun hashCode(): Int {
+        var result = hasReviewedAndConfirmedInfo.hashCode()
+        result = 31 * result + signature.contentHashCode()
+        return result
+    }
+}
